1 byte is enough for TelegramLogin.major

+++ src/main/java/io/github/giuliopft/telegramlogin/TelegramLogin.java
@@ -71,7 +71,7 @@ public final class TelegramLogin extends JavaPlugin {
     @Getter
     private final Map<Integer, Player> newPlayers = new ConcurrentHashMap<>();
     @Getter
-    private final short major = Short.parseShort(Bukkit.getBukkitVersion().split("\\.")[1]);
+    private final byte major = Byte.parseByte(Bukkit.getBukkitVersion().split("\\.")[1]);
 
     @Override
     public void onEnable() {
